Converted all UInt to Int, for now. Also made userDidLaunchApp() function public again for testing.

+++ SiriusRating/src/androidTest/java/com/theappcapital/siriusrating/ratingconditions/NotPostponedDueToReminderRatingConditionTest.kt
@@ -23,7 +23,7 @@ class NotPostponedDueToReminderRatingConditionTest {
 
     @Test
     fun test_condition_is_satisfied_when_the_user_did_not_opt_in_for_reminder() {
-        val remindMeLaterRatingCondition = NotPostponedDueToReminderRatingCondition(totalDaysBeforeReminding = 1u)
+        val remindMeLaterRatingCondition = NotPostponedDueToReminderRatingCondition(totalDaysBeforeReminding = 1)
 
         // By default the user did not opted-in for a reminder.
         // The condition should be satisfied, because the user did not opt-in for a reminder.
@@ -40,7 +40,7 @@ class NotPostponedDueToReminderRatingConditionTest {
     @Test
     fun test_condition_is_satisfied_when_the_user_opted_in_for_a_reminder_and_days_before_reminding_was_reached() {
         // Create the condition where we set the total days before reminding to 1.
-        val remindMeLaterRatingCondition = NotPostponedDueToReminderRatingCondition(totalDaysBeforeReminding = 1u)
+        val remindMeLaterRatingCondition = NotPostponedDueToReminderRatingCondition(totalDaysBeforeReminding = 1)
 
         // Set the action where the user did opt-in for a reminder yesterday.
         val dateTheUserOptedInForAReminder = LocalDateTime.now().minusDays(1).toInstant(ZoneOffset.UTC)
@@ -54,12 +54,12 @@ class NotPostponedDueToReminderRatingConditionTest {
     @Test
     fun test_condition_is_not_satisfied_when_the_user_opted_in_for_reminder_but_days_before_reminding_was_not_reached() {
         // Create the condition where we set the total days before reminding to 2.
-        val totalDaysBeforeReminding = 2u
+        val totalDaysBeforeReminding = 2
         val remindMeLaterRatingCondition = NotPostponedDueToReminderRatingCondition(totalDaysBeforeReminding = totalDaysBeforeReminding)
 
         // Create a random date that is 'now' or ((24 * totalDaysBeforeReminding) - 1) hours in the past.
         // For example: If the `totalDaysBeforeReminding` is 2, it will create a date in the past between 0 and 47 hours.
-        val dateTheUserOptedInForAReminder = LocalDateTime.now().minusHours((0 until (24 * totalDaysBeforeReminding.toInt() - 1)).random().toLong()).toInstant(ZoneOffset.UTC)
+        val dateTheUserOptedInForAReminder = LocalDateTime.now().minusHours((0 until (24 * totalDaysBeforeReminding - 1)).random().toLong()).toInstant(ZoneOffset.UTC)
         this.inMemorySiriusRatingDataStore.optedInForReminderUserActions = listOf(UserAction(appVersion = "0.1-anyversion", date = dateTheUserOptedInForAReminder))
 
         // The condition should not be satisfied, because the user opted-in for a reminder between 0 and 47 hours ago and
