+++ SiriusRating/src/main/java/com/theappcapital/siriusrating/ratingconditions/NotDeclinedToRateAnyVersionRatingCondition.kt
@@ -7,17 +7,10 @@ import java.time.Duration
 import java.time.LocalDate
 import java.time.ZoneOffset
 
-/**
- * The rating condition that checks if the user didn't decline the current version of the app.
- * We do not want to prompt the user to rate the app again if it declined to rate the current
- * version of the app.
- *
- * @property appVersionProvider
- */
 class NotDeclinedToRateAnyVersionRatingCondition(
-    private val daysAfterDecliningToPromptUserAgain: UInt,
+    private val daysAfterDecliningToPromptUserAgain: Int,
     private val backOffFactor: Double? = null,
-    private val maxRecurringPromptsAfterDeclining: UInt
+    private val maxRecurringPromptsAfterDeclining: Int
 ) : RatingCondition {
 
     /**
@@ -32,7 +25,7 @@ class NotDeclinedToRateAnyVersionRatingCondition(
         // If the the user didn't decline to rate the app (yet), return `true`.
         val mostRecentDeclinedToRateUserAction = declinedToRateUserActions.maxByOrNull { it.date } ?: return true
 
-        if ((declinedToRateUserActions.count() - 1) > this.maxRecurringPromptsAfterDeclining.toInt()) {
+        if ((declinedToRateUserActions.count() - 1) > this.maxRecurringPromptsAfterDeclining) {
             // We reached the maximum amount of times the user can decline, we do not want
             // to show the prompt anymore at this point, return `false`.
             return false
@@ -42,13 +35,13 @@ class NotDeclinedToRateAnyVersionRatingCondition(
         // 1. Get the total amount of days since the user last declined until now.
         val fromDate = LocalDate.from(mostRecentDeclinedToRateUserAction.date.atZone(ZoneOffset.UTC)).atStartOfDay()
         val nowDate = LocalDate.now(ZoneOffset.UTC).atStartOfDay()
-        val totalDaysAfterUserDeclinedLast = Duration.between(fromDate, nowDate).toDays().toUInt()
+        val totalDaysAfterUserDeclinedLast = Duration.between(fromDate, nowDate).toDays().toInt()
 
         // 2. Calculate the number of days it takes to show the prompt again after the user declined.
         val totalDaysAfterDecliningToPromptUserAgain = this.calculatedTotalDaysToPromptUserAgain(
             daysAfterDecliningToPromptUserAgain = this.daysAfterDecliningToPromptUserAgain,
             backOffFactor = this.backOffFactor,
-            timesDeclined = declinedToRateUserActions.count().toUInt()
+            timesDeclined = declinedToRateUserActions.count()
         )
 
         // Check if the total days after the user declined is greater than or equal
@@ -57,10 +50,10 @@ class NotDeclinedToRateAnyVersionRatingCondition(
     }
 
     fun calculatedTotalDaysToPromptUserAgain(
-        daysAfterDecliningToPromptUserAgain: UInt,
+        daysAfterDecliningToPromptUserAgain: Int,
         backOffFactor: Double?,
-        timesDeclined: UInt
-    ): UInt {
+        timesDeclined: Int
+    ): Int {
         // The `backOffFactor` must exist in order to make the back off calculation. If it does
         // not exist, return with the original amount of days.
         val backOffFactor = backOffFactor ?: return daysAfterDecliningToPromptUserAgain
@@ -68,7 +61,7 @@ class NotDeclinedToRateAnyVersionRatingCondition(
         // Formula: {days after declining to prompt user again} * ({back off factor} ^ ({times declined - 1}).
         // For example if the `daysAfterDecliningToPromptUserAgain` is '7' and `backOffFactor` is '2.0', it will
         // calculate [7-days, 14-days, 28-days, …] for the accumulating amount of declines.
-        return (daysAfterDecliningToPromptUserAgain.toDouble() * pow(backOffFactor, max(0, timesDeclined.toInt() - 1).toDouble())).toUInt()
+        return (daysAfterDecliningToPromptUserAgain.toDouble() * pow(backOffFactor, max(0, timesDeclined - 1).toDouble())).toInt()
     }
 
 }
